+++ src/main/java/fi/luke/bma/service/calculator/FreeformPolygonCalculator.java
@@ -56,7 +56,7 @@ public class FreeformPolygonCalculator extends SingleAreaCalculator {
         
         Map<String, String> displayOrders = new LinkedHashMap<>();
         for (Attribute attribute : sortedAttributes) {
-            displayOrders.put(Long.toString(attribute.getDisplayOrder()), attribute.getNameFI());   //TODO: Locale based
+            displayOrders.put(Double.toString(attribute.getDisplayOrder()), attribute.getNameFI());   //TODO: Locale based
         }
         result.put("displayOrders", displayOrders);
         return result;
